Moved components, deleted files
