--- src/common/c1-SuperInputText/AlternativeSuperInputText.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function AlternativeSuperInputText() {
-    return (
-        <input/>
-    )
-}
-
-export default AlternativeSuperInputText
