--- src/common/c1-SuperInputText/SuperInputText.module.css
@@ -1,23 +1,0 @@
-.superInput {
-    padding: 0 0 10px 0;
-    border: none; /*# Без рамки.*/
-    border-bottom: 2px solid #e0e0e0; /*# Нижняя рамка 1 пиксель, цвет  #e0e0e0*/
-    background-color: transparent; /*# Фон будет прозрачным.*/
-    outline: none;
-    transition: 0.3s;
-    color: #f7f8fa;
-}
-
-
-.errorInput {
-    padding: 0 0 10px 0;
-    border: none; /*# Без рамки.*/
-    border-bottom: 3px solid #e74c3c; /*# Нижняя рамка 1 пиксель, цвет  #e0e0e0*/
-    background-color: transparent; /*# Фон будет прозрачным.*/
-    outline: none;
-    transition: 0.3s;
-}
-
-.error {
-    color: #e74c3c;
-}
