--- src/common/c3-SuperCheckbox/AlternativeSuperCheckbox.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function AlternativeSuperCheckbox() {
-    return (
-        <input/>
-    )
-}
-
-export default AlternativeSuperCheckbox
