--- src/common/c3-SuperCheckbox/SuperCheckbox.module.css
@@ -1,7 +1,0 @@
-.checkbox {
-
-}
-
-.spanClassName {
-
-}
