+++ src/pages/Routes.tsx
@@ -4,7 +4,6 @@ import {Login} from "./auth/login/Login";
 import {Error404} from "./404/Error404";
 import {RestorePassword} from "./auth/recovery-password/restorePassword";
 import {SetNewPassword} from "./auth/enter-new-password/SetNewPassword";
-import {Example} from "./example/Example";
 import {Profile} from "./profile/Profile";
 import {RedirectPage} from "./auth/recovery-password/RedirectPage";
 import {Settings} from "./Settings";
@@ -19,7 +18,6 @@ export enum Path {
     error404 = '/error404',
     restorePassword = '/restore-password',
     setNewPassword = '/set-new-password',
-    example = '/example',
     redirectAfterSendRecoveryPassEmail = '/redirect-page',
     settings = '/settings',
     packsList = '/packsList',
@@ -40,7 +38,6 @@ export const Router = () => {
                 <Route path={Path.setNewPassword} element={<SetNewPassword/>}>
                     <Route path=':token' element={<SetNewPassword/>}/>
                 </Route>
-                <Route path={Path.example} element={<Example/>}/>
                 <Route path={Path.redirectAfterSendRecoveryPassEmail} element={<RedirectPage/>}/>
                 <Route path={Path.settings} element={<Settings/>}/>
                 <Route path={Path.packsList} element={<PacksListPage/>}/>
