--- src/pages/example/Example.tsx
@@ -1,11 +1,0 @@
-import SuperInputText from "../../common/c1-SuperInputText/SuperInputText";
-import SuperCheckbox from "../../common/c3-SuperCheckbox/SuperCheckbox";
-import SuperButton from "../../common/c2-SuperButton/SuperButton";
-
-export const Example = () => {
-    return <div>
-        <SuperInputText/>
-        <SuperCheckbox/>
-        <SuperButton>Example</SuperButton>
-    </div>
-}
